save spot and ww3forecaster

+++ surf-forecaster/gui/src/edu/unicen/surfforecaster/gwt/client/ForecastServicesAsync.java
@@ -6,5 +6,5 @@ import edu.unicen.surfforecaster.common.services.dto.PointDTO;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 
 public interface ForecastServicesAsync {
-	void getNearbyGridPoints(double spotLatitude, double spotLongitude, AsyncCallback<List<PointDTO>> callback);
+	void getNearbyGridPoints(float spotLatitude, float spotLongitude, AsyncCallback<List<PointDTO>> callback);
 }
